dev ver 0.3.47.1 1.修复了青雀规则的提示问题

+++ open_mahjong_unity/Assets/Scripts/GameScene/Calculation/Qingque13/Qingque13Calc/Core/QingqueTile.cs
@@ -75,7 +75,7 @@ namespace Qingque13.Core
         /// </summary>
         public QingqueSuit Suit(bool distinctDragons = false)
         {
-            return (QingqueSuit)(value & (byte)(0b11100000 + (distinctDragons ? 1 << 4 : 0)));
+            return (QingqueSuit)(value & (byte)(0b11100000 + (distinctDragons ? (byte)(1 << 4) : (byte)0)));
         }
 
         /// <summary>
